SetUp playerPrefs and added a folder for the Menu
made a script to load the player's preferences. The only thing left to set up  is for the ability to change key bindings and edit some of the functionality when it comes to brightness. I also decided to put all of the Main Menu scripts into their own folder for organization